+++ src/css/dev-tools.css
@@ -1,0 +1,49 @@
+.dev-tools-container {
+  position: fixed;
+  bottom: 20px;
+  left: 20px;
+  display: flex;
+  align-items: center;
+}
+
+.dev-main-button {
+  padding: 0;
+  background-color: transparent;
+  color: #fff;
+  border: none;
+  border-radius: 100%;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  font-size: 40px;
+}
+
+.dev-moving-button:hover {
+  margin-left: 10px;
+  opacity: 0;
+  transition: opacity 0.3s ease;
+  font-size: 20px;
+}
+
+.dev-button-bar {
+  display: none;
+  flex-direction: row;
+  align-items: center;
+  margin-left: 10px;
+  opacity: 0;
+  transition: opacity 0.3s ease;
+}
+
+.dev-button {
+  padding: 10px;
+  margin-right: 5px;
+  background-color: #f2f7fb;
+  color: #fff;
+  border: none;
+  border-radius: 20px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+
+.dev-button:hover {
+  background-color: #2980b9;
+}
