+++ src/css/modal-movie.css
@@ -1,5 +1,4 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap');
-
+/* kontener na okno modalne z informacjami o filmie */
 .modal {
   display: none;
   position: fixed;
@@ -12,21 +11,18 @@
   background-color: rgba(0, 0, 0, 0.5);
 }
 
+/* zawartość okna modalnego */
 .modal-content {
   background-color: #fefefe;
   margin: auto;
   border-radius: 15px;
   padding: 8px 20px 40px 20px;
   border: 1px solid #888;
-  max-width: 320px;
+  width: 320px;
   font-family: 'Roboto', sans-serif;
 }
 
-.modal-movie-info {
-  margin-top: 20px;
-  margin-left: auto;
-  max-width: 260px;
-}
+/* przycisk zamykania okna */
 .close {
   color: #aaaaaa;
   float: right;
@@ -35,6 +31,7 @@
   margin: 0;
 }
 
+/* wskazany przycisk zamykania okna */
 .close:hover,
 .close:focus {
   color: #000;
@@ -43,56 +40,27 @@
   scale: 1.2;
 }
 
-.dev-tools-container {
-  position: fixed;
-  bottom: 20px;
-  left: 20px;
-  display: flex;
-  align-items: center;
+/* zawartość okna modalnego z informacjami o filmie */
+.modal-container {
+  display: block;
 }
 
-.dev-main-button {
+/* plakat filmu w kontenerze */
+.movie-poster-modal img {
+  width: 100%;
+  height: 357px;
+  margin-top: 48px;
   padding: 0;
-  background-color: transparent;
-  color: #fff;
-  border: none;
-  border-radius: 100%;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  font-size: 40px;
+  margin: 0;
 }
 
-.dev-moving-button:hover {
-  margin-left: 10px;
-  opacity: 0;
-  transition: opacity 0.3s ease;
-  font-size: 20px;
+/* kontener na informacje o filmie */
+.modal-movie-info {
+  margin-top: 20px;
+  max-width: 260px;
 }
 
-.dev-button-bar {
-  display: none;
-  flex-direction: row;
-  align-items: center;
-  margin-left: 10px;
-  opacity: 0;
-  transition: opacity 0.3s ease;
-}
-
-.dev-button {
-  padding: 10px;
-  margin-right: 5px;
-  background-color: #f2f7fb;
-  color: #fff;
-  border: none;
-  border-radius: 20px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.dev-button:hover {
-  background-color: #2980b9;
-}
-
+/* tytuł filmu */
 .modal-movie-title {
   font-size: 20px;
   font-weight: 500;
@@ -101,6 +69,8 @@
   margin-bottom: 20px;
   overflow: hidden;
 }
+
+/* kontener na szczegółowe informacj eo filmie */
 .info-item {
   display: flex;
   justify-content: space-between;
@@ -109,6 +79,31 @@
   gap: 0px;
 }
 
+/* hasła szczegółowych informacji */
+.pernament-item {
+  font-size: 12px;
+  font-weight: 500;
+  line-height: 1.3;
+  min-width: 80px;
+  color: var(--vote-color);
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+}
+
+/* wartości szczegółowych informacji */
+.variables-item {
+  color: var(--movieDetails-color);
+  font-size: 12px;
+  font-weight: 500;
+  line-height: 1.3;
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+  padding-right: 30px;
+}
+
+/* średnia nota */
 .average-vote {
   background-color: var(--details-contrast);
   border-radius: 5px;
@@ -117,6 +112,7 @@
   margin-right: 4px;
 }
 
+/* liczba głosów */
 .count-vote {
   background-color: var(--footer-bg);
   border-radius: 5px;
@@ -128,29 +124,13 @@
 .slash-color {
   color: var(--vote-color);
 }
-.pernament-item {
-  font-size: 12px;
-  font-weight: 500;
-  line-height: 1.3;
-  min-width: 80px;
-  color: var(--vote-color);
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-}
-.variables-item {
-  color: var(--movieDetails-color);
-  font-size: 12px;
-  font-weight: 500;
-  line-height: 1.3;
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-  padding-right: 30px;
-}
+
+/* oryginalny tytuł */
 .original-title {
   text-transform: uppercase;
 }
+
+/* opis filmu */
 .about-movie {
   font-size: 12px;
   font-weight: 500;
@@ -160,6 +140,12 @@
   margin-top: 20px;
 }
 
+/* hasło przed opisem filmu */
+.about-movie-details {
+  text-transform: uppercase;
+}
+
+/* kontener na przyciski dodawania filmu do listy obejrzanych lub do obejrzenia */
 .modal-buttons {
   margin-top: 20px;
   margin-bottom: 20px;
@@ -167,10 +153,7 @@
   justify-content: space-between;
 }
 
-.modal-buttons button {
-  gap: 20px;
-}
-
+/* kontener na przycisk do włączenia trailera filmu */
 .movie-trailer {
   display: flex;
   flex-direction: column;
@@ -178,22 +161,9 @@
   align-items: center;
 }
 
-.movie-trailer button {
-  border-radius: 5px;
-  width: 110px;
-  height: 44px;
-  color: var(--bg-color);
-  cursor: pointer;
-  border: none;
-  color: var(--bg-color);
-  font-weight: 500;
-  font-size: 12px;
-  line-height: 1.3;
-  text-transform: uppercase;
-  text-align: center;
-}
-.watchedButton,
-.queuedButton {
+/* przyciski dodawania filmu do listy obejrzanych lub do obejrzenia */
+.watched-button,
+.queued-button {
   border-radius: 5px;
   width: 110px;
   height: 44px;
@@ -210,40 +180,41 @@
   padding: 6px 27px;
 }
 
-.watchedButton:hover,
-.watchedButton:focus,
-.queuedButton:hover,
-.queuedButton:focus,
+/* przycisk do włączenia trailera filmu */
+.trailer-button {
+  border-radius: 5px;
+  width: 110px;
+  height: 44px;
+  color: var(--bg-color);
+  cursor: pointer;
+  border: none;
+  color: var(--bg-color);
+  font-weight: 500;
+  font-size: 12px;
+  line-height: 1.3;
+  text-transform: uppercase;
+  text-align: center;
+}
+
+/* wskazane przyciski */
+.watched-button:hover,
+.watched-button:focus,
+.queued-button:hover,
+.queued-button:focus,
 .movie-trailer button:hover,
 .movie-trailer button:focus {
   transform: scale(1.1);
 }
 
-.about-movie-details {
-  text-transform: uppercase;
-}
-
-.movie-poster-modal img {
-  width: 100%;
-  height: 357px;
-  margin-top: 48px;
-  padding: 0;
-  margin: 0;
-}
-
-.modal-container {
-  display: block;
-}
-
+/* style dla tabletu i komputera */
 @media screen and (min-width: 768px) {
   .modal-content {
-    max-width: 704px;
+    width: 704px;
     min-height: 552px;
   }
   .modal-container {
     display: flex;
   }
-
   .movie-poster-modal {
     margin-right: 68px;
     margin-top: 40px;
@@ -254,10 +225,8 @@
     margin: 0;
     padding: 0;
   }
-
   .modal-movie-info {
     margin-top: 41px;
-    margin-left: -41px;
   }
   .info-item {
     margin-top: 20px;
@@ -266,13 +235,14 @@
 
 @media (min-width: 1280px) {
   .modal-content {
+    width: 806px;
     min-height: 558px;
     padding: 0 20px;
   }
   .modal-container {
     margin-bottom: 30px;
   }
-  .movieData.title {
+  .modal-movie-title {
     font-size: 30px;
   }
   .movie-poster-modal {
@@ -284,10 +254,9 @@
   }
   .modal-movie-info {
     margin-top: 41px;
-    margin-left: auto;
   }
   .about-movie {
-    max-width: 391px;
+    width: 391px;
   }
   .variables-item {
     padding-right: 0;
