+++ src/index.html
@@ -17,9 +17,10 @@
     <include src="./partials/movies.html"></include>
     <include src="./partials/pagination.html"></include>
     <include src="./partials/footer.html"></include>
-    <include src="./partials/modal.html"></include>
+    <include src="./partials/modal-movie.html"></include>
+    <include src="./partials/dev-tools.html"></include>
     <include src="./partials/modal-goit-team.html"></include>
-    
+    <include src="./partials/registration.html"></include>
 
     <script src="./index.js" type="module"></script>
     <script src="https://unpkg.com/gsap@3.9.1/dist/gsap.min.js"></script>
