+++ src/index.js
@@ -2,13 +2,11 @@
 import './sass/main.scss';
 //JS Global
 import './js/api';
-//import './js/search';
 import './js/gallery';
-import './js/markup';
 import './js/localstorage';
-import './js/visibilityHeader';
-import './js/onOffModal';
-import './js/devTools';
-import './js/darkMode';
+import './js/header';
+import './js/modal-movie';
+import './js/dev-tools';
+import './js/dark-mode';
 import './js/modal-goit-team';
 import './js/firebaseAuthorization';
