+++ src/js/gallery.js
@@ -1,10 +1,9 @@
-// gallery.js
-
 import {
   fetchMovieDetails,
   fetchMovieTrailers,
   fetchSearchMovies,
   fetchTrendingMovies,
+  fetchGenres,
   genresName,
 } from './api';
 import { addToQueue, addToWatchedMovies } from './localstorage';
@@ -17,17 +16,18 @@ let isInfinityScrollEnable = 0;
 let searchQuery;
 let totalPages;
 
-// Funkcja pomocnicza do pobrania nazw gatunków na podstawie ich identyfikatorów
+// tworzenie napisu z nazwami gatunków filmu na podstawie ich identyfikatorów
 export const getGenres = genreIds => {
-  // Pobranie nazw gatunków z listy genresName zdefiniowanej w api.js
+  // pobranie psaujących nazw gatunków z listy genresName zdefiniowanej w api.js
   const genres = genreIds.map(genreId => {
     const foundGenre = genresName.find(genre => genre.id === genreId);
     return foundGenre ? foundGenre.name : '';
   });
-  // Zwrócenie połączonej listy gatunków
+  // zwrócenie połączonej listy gatunków
   return genres.join(', ');
 };
 
+// wyświetlenie galerii filmów z listy obejrzanych
 const displayWatchedMovies = () => {
   try {
     const watchedMovies = JSON.parse(localStorage.getItem('watchedMovies')) || [];
@@ -47,6 +47,7 @@ const displayWatchedMovies = () => {
   }
 };
 
+// wyświetlenie galerii filmów z listy do obejrzenia
 const displayQueuedMovies = () => {
   try {
     const queuedMovies = JSON.parse(localStorage.getItem('queuedMovies')) || [];
@@ -66,12 +67,12 @@ const displayQueuedMovies = () => {
   }
 };
 
-const displayMovieDetails = movieDetails => {
-  // Tutaj możemy zaimplementować logikę wyświetlania informacji o filmie w modalu
-  // console.log(movieDetails);
-};
+// const displayMovieDetails = movieDetails => {
+//   // Tutaj możemy zaimplementować logikę wyświetlania informacji o filmie w modalu
+//   // console.log(movieDetails);
+// };
 
-//Obsługa HomePage i Buttonów
+// obsługa zdarzenia kliknięcia przycisków HomePage i Buttonów
 window.addEventListener('DOMContentLoaded', () => {
   getHomepage(1); // Wywołujemy funkcję wyświetlającą HomePage
 
@@ -88,7 +89,7 @@ window.addEventListener('DOMContentLoaded', () => {
   // libraryQueuedButton.addEventListener('click', displayQueuedMovies);
 });
 
-//Generujemy trendings movie
+// wyświetlanie podanej strony galerii popularnych filmów
 export const getHomepage = async (pageNo, infinity) => {
   try {
     const response = await fetchTrendingMovies(pageNo);
@@ -105,7 +106,7 @@ export const getHomepage = async (pageNo, infinity) => {
   }
 };
 
-//Obsługa szukajki
+// obsługa zdarzenia kliknięcia formularza wyszukiwania
 document.addEventListener('DOMContentLoaded', () => {
   const searchForm = document.getElementById('search-form');
   const searchInput = document.querySelector('.search-form input');
@@ -117,6 +118,7 @@ document.addEventListener('DOMContentLoaded', () => {
   });
 });
 
+// wyświetlanie podanej strony galerii szukanych filmów
 export const getSearchResult = async (event, pageNo) => {
   event.preventDefault();
   homePageNo = pageNo;
@@ -349,27 +351,27 @@ const openModal = movieData => {
       </div>
 
       <div class="modal-buttons">
-        <button class="watchedButton">Add to Watched</button>
-        <button class="queuedButton">Add to Queue</button>
+        <button id="watchedButton" class="watched-button">Add to Watched</button>
+        <button id="queueButton" class="queued-button">Add to Queue</button>
 
       </div>
       <div class="movie-trailer">
-      <button id="movieTrailerButton">Trailer</button>
+      <button id="movieTrailerButton" class="trailer-button">Trailer</button>
       </div>
     </div>
   </div>
   `;
 
-  const watchedButton = document.getElementsByClassName('watchedButton')[0];
+  const watchedButton = document.getElementsById('watchedButton')[0];
   watchedButton.onclick = () => {
     addToWatchedMovies(movieData);
   };
-  const queuedButton = document.getElementsByClassName('queuedButton')[0];
+  const queuedButton = document.getElementsById('queuedButton')[0];
   queuedButton.onclick = () => {
     addToQueue(movieData);
   };
 
-  // const span = document.getElementsByClassName('close-modal-movie')[0];
+  // const span = document.getElementsById('close-modal-movie')[0];
   // span.onclick = () => {
   //   modal.style.display = 'none';
   // };
@@ -438,13 +440,14 @@ const openModal = movieData => {
     modal.style.display = 'none';
   };
 
-  // Obsługa zdarzenia keydown
+  // obsługa zdarzenia naciśnięcia klawisza Escape do wyłączenia okna modalnego
   document.addEventListener('keydown', function (event) {
     if (event.key === 'Escape') {
       modal.style.display = 'none';
     }
   });
 
+  // obsługa zdarzenia kliknięcia do wyłączenia okna modalnego
   window.onclick = event => {
     if (event.target == modal) {
       modal.style.display = 'none';
@@ -491,11 +494,11 @@ const openModal = movieData => {
 //     </div>
 //     </div>
 //   `;
-//   const watchedButton = document.getElementsByClassName('watchedButton')[0];
+//   const watchedButton = document.getElementsById('watchedButton')[0];
 //   watchedButton.onclick = () => {
 //     addToWatchedMovies(movieData);
 //   };
-//   const queuedButton = document.getElementsByClassName('queuedButton')[0];
+//   const queuedButton = document.getElementsById('queuedButton')[0];
 //   queuedButton.onclick = () => {
 //     addToQueue(movieData);
 //   };
@@ -552,7 +555,7 @@ const openModal = movieData => {
 //     }
 //   });
 
-//   const span = document.getElementsByClassName('close')[0];
+//   const span = document.getElementsById('close')[0];
 //   span.onclick = () => {
 //     modal.style.display = 'none';
 //   };
@@ -620,7 +623,7 @@ const loadMoreContent = () => {
 };
 const infinityScroll = document.getElementById('infinityScroll');
 
-// Obsługa zdarzenia kliknięcia przycisku
+// obsługa zdarzenia kliknięcia przycisku automatycznego pobierania kolejnych stron w miarę przewijania
 infinityScroll.addEventListener('click', () => {
   if (isInfinityScrollEnable) {
     // Jeżeli infinity scroll jest aktywny, usuwamy nasłuchiwanie zdarzenia scroll
