+++ src/partials/dev-tools.html
@@ -1,0 +1,9 @@
+<div class="dev-tools-container">
+  <button class="dev-main-button" id="dev-mainButton">⚙️</button>
+  <div class="dev-button-bar">
+    <button class="dev-button" id="changeMode">🌗</button>
+    <button class="dev-button" id="scrollToTop">🔝</button>
+    <button class="dev-button" id="infinityScroll">♾️</button>
+    <button class="dev-button" id="wide-container">↔️</button>
+  </div>
+</div>
