+++ src/partials/modal-movie.html
@@ -1,0 +1,11 @@
+<!-- Aleksander -->
+<div id="myModal" class="modal">
+  <div class="modal-content">
+    <span class="close">&times;</span>
+    <div id="modalContent">
+      <!-- Tutaj będą wyświetlane dynamiczne dane -->
+    </div>
+  </div>
+</div>
+
+
