+++ src/sass/main.scss
@@ -10,6 +10,7 @@
 @import '../css/movies.css';
 @import '../css/pagination.css';
 @import '../css/footer.css';
+@import '../css/dev-tools.css';
 @import '../css/modal-movie.css';
 @import '../css/modal-goit-team.css';
 @import '../css/modal-registration.css';
